localstorage for token
Localstorage token logic added to project to save token for localstorage, also pings server on every fetch to check if user is logged in and in a case of unauthorized it's jumps to login page.

Work needed:
due to useffect the page still have the auth state. This needs to be changed in the future!